use double, not float

+++ pilot/modules/protocol/protocol.c
@@ -196,14 +196,14 @@ double protocol_byte_to_radian(uint8_t x) {
 	double r = 0.024639942381096 * x;
 	return (r > M_PI) ? 2.0 * M_PI - r : r; // convert quadrants III & IV into negative values
 }
-uint8_t protocol_radian_to_byte(float x) {
+uint8_t protocol_radian_to_byte(double x) {
 	double r = x < 0 ? 2.0 * M_PI - x : x; // convert quadrants III & IV into positive values
 	return 40.584510488433314 * r;
 }
 double protocol_byte_to_percent(uint8_t x) {
 	return 0.392156862745098 * x;
 }
-uint8_t protocol_percent_to_byte(float x) {
+uint8_t protocol_percent_to_byte(double x) {
 	return 2.55 * x;
 }
 void protocol_double_to_bytes(double value, uint8_t *buffer) {
